refactor(trainer): move _dataset.py into a _private/ subpackage
Matches the internal-module convention used elsewhere in the repo
(lib/trainer/torch/pytorch_lightning/_private/, lib/model_manager/_private/)
and now also used by _tracker.py in this same package (PR #1432). Pure
rename + import path updates; no behavior change.
